pythia run scripts

+++ tasks/evaluator.py
@@ -167,7 +167,7 @@ class TaskEvaluator:
             model_path,
             revision=self.model_config.checkpoint,
             trust_remote_code=self.model_config.trust_remote_code,
-            torch_dtype=torch.float16,
+            torch_dtype=torch.bfloat16,  # avoid float16 overflow (NaN logits) at early checkpoints
             device_map="auto"
         )
